Improve: 신고 이력 map 집계 쿼리 최적화 2.5 s -> 1.3 s

+++ api-module/src/main/kotlin/com/cj/smartworker/business/fcm/dto/response/HeartRateAggregateResponse.kt
@@ -1,7 +1,7 @@
 package com.cj.smartworker.business.fcm.dto.response
 
 data class HeartRateAggregateResponse(
-    val x: Float,
-    val y: Float,
+    val x: Double,
+    val y: Double,
     val count: Long,
 )
